+++ msx2cas-cmd/src/br/com/dod/vcas/ConvertFile.java
@@ -1,18 +1,18 @@
 package br.com.dod.vcas;
 
-public class ConvertFile {	
+class ConvertFile {
 	private String inputName;
 	private String outputName;
 	
-	public ConvertFile(String inputName, String outputName) {
+	ConvertFile(String inputName, String outputName) {
 		this.inputName = inputName;
 		this.outputName = outputName;
 	}
 	
-	public String getInputName() {
+	String getInputName() {
 		return inputName;
 	}
-	public String getOutputName() {
+	String getOutputName() {
 		return outputName;
 	}
 }
