+++ msx2cas-core/src/main/java/br/com/dod/vcas/util/FileCommons.java
@@ -1,10 +1,11 @@
-package br.com.dod.vcas;
+package br.com.dod.vcas.util;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 
 import br.com.dod.vcas.exception.FlowException;
+import br.com.dod.vcas.model.FileType;
 
 public class FileCommons {
 
@@ -21,13 +22,10 @@ public class FileCommons {
 
 	public static byte[] readFile(File inputFile) throws FlowException {
 		byte[] inputMemPointer;
-		FileInputStream fis = null;
-		try {
+		try (final FileInputStream fis = new FileInputStream(inputFile)) {
 			if (!inputFile.exists()) throw FlowException.error("file_not_found");
 			if (!inputFile.isFile()) throw FlowException.error("not_a_file");
 			if (!inputFile.canRead()) throw FlowException.error("file_access_denied");
-
-			fis = new FileInputStream(inputFile);
 
 			inputMemPointer = new byte[(int) inputFile.length()];
 
@@ -36,17 +34,11 @@ public class FileCommons {
 			}
 		} catch (IOException e) {
 			throw FlowException.error("file_not_found");
-		} finally {
-			try {
-				if (fis != null) fis.close();
-			} catch (IOException e) {
-				throw FlowException.error("file_access_denied");
-			}
 		}
 		return inputMemPointer;
 	}
 
-	static FileType detectFile(String fileName) throws IOException {
+	public static FileType detectFile(String fileName) throws IOException {
 		FileType fileType = FileType.ASCII;
 
 		if (FileType.CAS.equals(fileName)) {
@@ -63,7 +55,7 @@ public class FileCommons {
 
 	}
 
-	static FileType detectFile(byte[] inputHandler) throws IOException {
+	public static FileType detectFile(byte[] inputHandler) throws IOException {
 		FileType fileType = FileType.ASCII;
 
 		if (FileType.CAS.equals(inputHandler)) {
