+++ msx2cas-core/src/main/java/br/com/dod/vcas/util/WavHeader.java
@@ -1,4 +1,4 @@
-package br.com.dod.vcas;
+package br.com.dod.vcas.util;
 
 import java.util.LinkedList;
 import java.util.List;
